+++ packages/menu/src/menu.vue
@@ -1,78 +1,79 @@
 <template>
-  <el-menu
-    :mode="mode"
-    :collapse="collapse"
-    :background-color="backgroundColor"
-    :text-color="textColor"
-    :active-text-color="activeTextColor"
-    :default-active="defaultActive"
-    :default-openeds="defaultOpeneds"
-    :unique-opened="uniqueOpened"
-    :menu-trigger="menuTrigger"
-    :router="router"
-    :collapse-transition="collapseTransition"
-    :style="{ width: this.width }"
-    v-on="$listeners"
-  >
-    <template v-for="(menuItem, menuIndex) of menuList">
-      <!-- [20210430][upg] 存在下级菜单 -->
-      <template v-if="menuItem[listSetting.children || 'children']">
-      <el-submenu
-        :key="menuIndex"
-        :index="`${menuItem[listSetting.index || 'index']}`"
-        :data-index="`${menuItem[listSetting.index || 'index']}`"
-        :disabled="menuItem[listSetting.disabled || 'disabled']"
-      >
-        <template slot="title">
-          <i class="el-icon-location"></i>
-            <!-- <router-link :to="menuItem.url"> -->
-            <template v-if="menuItem[listSetting.href || 'href']">
-              <a
-                class="a-underline-none"
-                :href="menuItem[listSetting.href || 'href']"
-                target="_blank"
-              >
-                {{ menuItem[listSetting.menuName || 'menuName'] }}
-              </a>
-            </template>
-
-            <template v-else>
-              {{ menuItem[listSetting.menuName || 'menuName'] }}
-            </template>
-            <!-- </router-link> -->
-        </template>
-        <template v-for="(subMenuItem, subMenuIndex) of menuItem[listSetting.children || 'children']">
-          <!-- id：用于按钮权限控制，不添加影响部分列表页添加按钮功能 -->
-          <!-- <router-link :to="`${subMenuItem.label}?id=${subMenuItem.id}`"> -->
-          <el-menu-item
-            :key="subMenuIndex"
-            :index="`${menuItem[listSetting.index || 'index']}-${subMenuItem[listSetting.index || 'index']}-${subMenuIndex}`"
-            :data-index="`${menuItem[listSetting.index || 'index']}-${subMenuItem[listSetting.index || 'index']}-${subMenuIndex}`"
-            :disabled="subMenuItem[listSetting.disabled || 'disabled']"
-            @click="$emit('menu-item-click', subMenuItem)"
+  <div class="es-menu">
+    <el-menu
+      :mode="mode"
+      :collapse="collapse"
+      :background-color="backgroundColor"
+      :text-color="textColor"
+      :active-text-color="activeTextColor"
+      :default-active="defaultActive"
+      :default-openeds="defaultOpeneds"
+      :unique-opened="uniqueOpened"
+      :menu-trigger="menuTrigger"
+      :router="router"
+      :collapse-transition="collapseTransition"
+      :style="{ width: this.width }"
+      v-on="$listeners"
+    >
+      <template v-for="(menuItem, menuIndex) of menuList">
+        <!-- [20210430][upg] 存在下级菜单 -->
+        <template v-if="menuItem[listSetting.children || 'children']">
+          <el-submenu
+            :key="menuIndex"
+            :index="`${menuItem[listSetting.index || 'index']}`"
+            :data-index="`${menuItem[listSetting.index || 'index']}`"
+            :disabled="menuItem[listSetting.disabled || 'disabled']"
           >
-            <template v-if="subMenuItem[listSetting.href || 'href']">
-              <i class="iconfont"></i>
-              <a
-                :href="subMenuItem[listSetting.href || 'href']"
-                target="_blank"
-              >
-                {{ subMenuItem[listSetting.menuName || 'menuName'] }}
-              </a>
-            </template>
+            <template slot="title">
+              <i class="el-icon-location"></i>
+              <!-- <router-link :to="menuItem.url"> -->
+              <template v-if="menuItem[listSetting.href || 'href']">
+                <a
+                  class="a-underline-none"
+                  :href="menuItem[listSetting.href || 'href']"
+                  target="_blank"
+                >
+                  {{ menuItem[listSetting.menuName || 'menuName'] }}
+                </a>
+              </template>
 
-            <template v-else>
-              <i class="iconfont"></i>
-              {{ subMenuItem[listSetting.menuName || 'menuName'] }}
+              <template v-else>
+                {{ menuItem[listSetting.menuName || 'menuName'] }}
+              </template>
+              <!-- </router-link> -->
             </template>
-          </el-menu-item>
-          <!-- </router-link> -->
+            <template v-for="(subMenuItem, subMenuIndex) of menuItem[listSetting.children || 'children']">
+              <!-- id：用于按钮权限控制，不添加影响部分列表页添加按钮功能 -->
+              <!-- <router-link :to="`${subMenuItem.label}?id=${subMenuItem.id}`"> -->
+              <el-menu-item
+                :key="subMenuIndex"
+                :index="`${menuItem[listSetting.index || 'index']}-${subMenuItem[listSetting.index || 'index']}-${subMenuIndex}`"
+                :data-index="`${menuItem[listSetting.index || 'index']}-${subMenuItem[listSetting.index || 'index']}-${subMenuIndex}`"
+                :disabled="subMenuItem[listSetting.disabled || 'disabled']"
+                @click="$emit('menu-item-click', subMenuItem)"
+              >
+                <template v-if="subMenuItem[listSetting.href || 'href']">
+                  <i class="iconfont"></i>
+                  <a
+                    :href="subMenuItem[listSetting.href || 'href']"
+                    target="_blank"
+                  >
+                    {{ subMenuItem[listSetting.menuName || 'menuName'] }}
+                  </a>
+                </template>
+
+                <template v-else>
+                  <i class="iconfont"></i>
+                  {{ subMenuItem[listSetting.menuName || 'menuName'] }}
+                </template>
+              </el-menu-item>
+              <!-- </router-link> -->
+            </template>
+          </el-submenu>
         </template>
-      </el-submenu>
-      </template>
-      <!-- [20210430][upg] 不存在下级菜单 -->
-      <template v-else>
-        <el-menu-item
+        <!-- [20210430][upg] 不存在下级菜单 -->
+        <template v-else>
+          <el-menu-item
             :key="menuIndex"
             :index="`${menuItem[listSetting.index || 'index']}`"
             :data-index="`${menuItem[listSetting.index || 'index']}`"
@@ -95,9 +96,10 @@
               {{ menuItem[listSetting.menuName || 'menuName'] }}
             </template>
           </el-menu-item>
+        </template>
       </template>
-    </template>
-  </el-menu>
+    </el-menu>
+  </div>
 </template>
 <script>
 /**
@@ -122,15 +124,15 @@ export default {
     // 导航菜单数组组成配置
     listSetting: {
       type: Object,
-      default: ()=>{
+      default: () => {
         return {
           index: 'index',
           menuName: 'menuName',
           children: 'children',
           disabled: 'disabled',
-          href: 'href'
+          href: 'href',
         }
-      }
+      },
     },
     // 导航栏宽度
     width: {
@@ -141,7 +143,7 @@ export default {
 }
 </script>
 <style>
-  a {
-    text-decoration: none;
-  }
+a {
+  text-decoration: none;
+}
 </style>