+++ CHANGELOG.md
@@ -36,3 +36,44 @@
 # [1.0.14][20210428]
 + [upd] <es-menu>
 - 修复子菜单点击触发实参非当前点击子菜单对象问题
+
+# [1.0.15][20210506]
++ [crt] 添加 <es-tabs> 标签页
+- 继承 <el-tabs>
+- 通过新增属性[list-setting] 设置传入标签页数组[tab-list] 中对象属性值
+```javascript
+// [tab-list] 数组示例
+[
+  { label: "标签一", name: "标签一" },
+  { label: "标签二", name: "标签二" }
+]
+```
++ [crt] 添加 <es-crumb> 面包屑
+- 继承 <es-breadCrumb>
+- 通过新增属性[list-setting] 设置传入标签页数组[crumb-list] 中对象属性值
+- 通过新增属性[button-list-setting] 设置按钮组[button-list] 对象属性值
+```javascript
+// [crumb-list] 数组示例
+[
+  { name: "标签一" },
+  { name: "标签二" }
+]
+```
++ [crt] 添加 <es-button> 按钮
+- 继承 <es-button>
+
++ [crt] 添加 <es-button-group> 按钮组
+- 通过属性[list-setting] 设置传入标签页数组[button-list] 中对象属性值
+```javascript
+// [button-list] 数组示例
+[
+  { buttonName: '新增企业', size: 'mini', type: 'success', clickEvent: 'addCompany' },
+  { buttonName: '新增部门', size: 'mini', type: 'warning', clickEvent: 'addDept' },
+  { buttonName: '新增员工', size: 'mini', type: 'danger', visible: false }
+]
+```
++ [crt] 添加 <es-query> 条件查询组件
+- 通过[query-list] 配置查询条件表单
+- 通过[button-list] 配置查询/重置按钮，默认自带
+
+